+++ src/main/java/site/scalarstudios/scalarpower/block/ScalarPowerBlocks.java
@@ -31,6 +31,7 @@ import site.scalarstudios.scalarpower.machines.wire.glassfiber.GlassFiberWireBlo
 import site.scalarstudios.scalarpower.machines.wire.gold.GoldWireBlock;
 import site.scalarstudios.scalarpower.machines.wire.gold.InsulatedGoldWireBlock;
 import site.scalarstudios.scalarpower.machines.wire.reinforcedglassfiber.ReinforcedGlassFiberWireBlock;
+import site.scalarstudios.scalarpower.machines.redstoneclock.RedstoneClockBlock;
 import site.scalarstudios.scalarpower.item.ScalarPowerItems;
 import site.scalarstudios.scalarpower.item.custom.TooltipBlockItem;
 
@@ -134,6 +135,11 @@ public class ScalarPowerBlocks {
     public static final DeferredBlock<ReinforcedGlassFiberWireBlock> REINFORCED_GLASS_FIBER_WIRE = registerTooltipBlockItem("reinforced_glass_fiber_wire",
             ReinforcedGlassFiberWireBlock::new,
             properties -> properties.strength(0.5F, 0.5F).noOcclusion().sound(SoundType.GLASS).mapColor(MapColor.COLOR_LIGHT_BLUE));
+
+    /* Redstone Utilities */
+    public static final DeferredBlock<RedstoneClockBlock> REDSTONE_CLOCK = registerBlock("redstone_clock",
+            RedstoneClockBlock::new,
+            properties -> properties.strength(2.5F, 2.5F).requiresCorrectToolForDrops().sound(SoundType.COPPER).mapColor(MapColor.COLOR_ORANGE));
 
     // Normal Blocks
     public static final DeferredBlock<Block> COBALT_BLOCK = registerBlock("cobalt_block",